chore: clean up sidebar and settings icons and move them to a separate util
